preprocessing/requirements_pre.txt
numpy
scipy
soundfile
tqdm
torch
torchaudio
google-cloud-storage
transformers
huggingface_hub
einops
librosa
nnAudio
easydict
x-clip

# MuQ (pip-installable package)
git+https://github.com/tencent-ailab/MuQ.git
